backend/requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-DebugToolbar==0.11.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-dotenv==0.16.0
Werkzeug==1.0.1
flask-cors==4.0.0
faker==21.0.0
openai==1.4.0